+++ app/src/main/java/com/example/interpark/adapters/ReviewAdapter.kt
@@ -3,13 +3,10 @@ package com.example.interpark.adapters
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
 import android.widget.RatingBar
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import coil.load
 import com.example.interpark.R
-import com.example.interpark.data.Performance
 import com.example.interpark.data.types.Review
