+++ app/src/main/java/com/example/interpark/adapters/SeatAdapter.kt
@@ -4,10 +4,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.example.interpark.R
-import com.example.interpark.data.Seat
+import com.example.interpark.data.types.Seat
 
 class SeatAdapter(
     private val seatList: List<Seat>,
