--- app/src/main/java/com/example/interpark/data/login.kt
@@ -1,45 +1,0 @@
-package com.example.interpark.data
-
-//import retrofit2.Call
-//import retrofit2.http.Body
-//import retrofit2.http.GET
-//import retrofit2.http.Header
-//import retrofit2.http.POST
-//
-//// 요청에 필요한 데이터 클래스 정의
-//data class SignUpRequest(
-//    val username: String,
-//    val password: String,
-//    val nickname: String,
-//    val phoneNumber: String,
-//    val email: String
-//)
-//
-//data class User(
-//    val id: String,
-//    val username: String,
-//    val nickname: String,
-//    val phoneNumber: String,
-//    val email: String
-//)
-//
-//data class SignUpResponse(
-//    val user: User
-//)
-//
-//data class SignInRequest(
-//    val username: String,
-//    val password: String
-//)
-//
-//data class SignInResponse(
-//    val accessToken: String
-//)
-//
-//data class UserProfile(
-//    val username: String,
-//    val nickname: String,
-//    val phoneNumber: String,
-//    val email: String
-//)
-//
