+++ app/src/main/java/com/example/interpark/data/types/Seat.kt
@@ -1,4 +1,4 @@
-package com.example.interpark.data
+package com.example.interpark.data.types
 
 data class Seat(
     val row: String, // 좌석의 행 정보 (예: "A", "B")
