+++ app/src/main/java/com/example/interpark/fragments/CategoryPage/CategoryFragmentHome.kt
@@ -1,12 +1,12 @@
-package com.example.interpark.fragments
+package com.example.interpark.fragments.CategoryPage
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
 import com.example.interpark.R
 
-class SearchFragmentHome: Fragment(R.layout.fragment_search_home) {
+class CategoryFragmentHome : Fragment(R.layout.fragment_category_home){
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-    super.onViewCreated(view, savedInstanceState)
-}
+        super.onViewCreated(view, savedInstanceState)
+    }
 }
