+++ app/src/main/java/com/example/interpark/fragments/CategoryPage/EmptyFragment.kt
@@ -1,7 +1,6 @@
-package com.example.interpark.fragments
+package com.example.interpark.fragments.CategoryPage
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -9,7 +8,6 @@ import android.widget.ImageButton
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.findNavController
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
@@ -17,7 +15,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.interpark.R
 import com.example.interpark.adapters.PerformanceAdapter
-import com.example.interpark.data.Performance
+import com.example.interpark.data.types.Performance
 import com.example.interpark.databinding.FragmentEmptyBinding
 import com.example.interpark.viewModels.MyPageViewModel
 import com.example.interpark.viewModels.MyPageViewModelFactory
