--- app/src/main/java/com/example/interpark/fragments/MyFragmentHome.kt
@@ -1,13 +1,0 @@
-package com.example.interpark.fragments
-
-import android.os.Bundle
-import android.view.View
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.NavHostFragment
-import com.example.interpark.R
-
-class MyFragmentHome : Fragment(R.layout.fragment_my_home){
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-}
