+++ app/src/main/java/com/example/interpark/fragments/MyPage/MyFragmentAccount.kt
@@ -1,4 +1,4 @@
-package com.example.interpark.fragments
+package com.example.interpark.fragments.MyPage
 
 import android.os.Bundle
 import android.view.LayoutInflater
