+++ app/src/main/java/com/example/interpark/fragments/MyPage/MyFragmentHome.kt
@@ -1,12 +1,11 @@
-package com.example.interpark.fragments
+package com.example.interpark.fragments.MyPage
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.NavHostFragment
 import com.example.interpark.R
 
-class CategoryFragmentHome : Fragment(R.layout.fragment_category_home){
+class MyFragmentHome : Fragment(R.layout.fragment_my_home){
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
     }
