--- app/src/main/java/com/example/interpark/fragments/PaymentFragment.kt
@@ -1,4 +1,0 @@
-package com.example.interpark.fragments
-
-class PaymentFragment {
-}
