+++ app/src/main/java/com/example/interpark/fragments/PerformanceDetail/Reviews.kt
@@ -1,4 +1,4 @@
-package com.example.interpark.fragments.detailFragments
+package com.example.interpark.fragments.PerformanceDetail
 
 import android.os.Build
 import android.os.Bundle
@@ -6,27 +6,16 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import android.widget.ImageView
-import android.widget.LinearLayout
-import android.widget.TextView
 import androidx.annotation.RequiresApi
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
 import com.example.interpark.R
 import com.example.interpark.adapters.ReviewAdapter
 import com.example.interpark.data.types.Review
-import com.example.interpark.databinding.FragmentPerformanceDetailBinding
-import com.example.interpark.databinding.FragmentPerformanceDetailPerformanceInformationBinding
 import com.example.interpark.databinding.FragmentPerformanceDetailReviewsBinding
-import com.example.interpark.fragments.PerformanceDetailFragmentArgs
-import com.example.interpark.fragments.PerformanceDetailFragmentDirections
 import com.example.interpark.viewModels.PerformanceDetailViewModel
 import com.example.interpark.viewModels.PerformanceDetailViewModelFactory
