+++ app/src/main/java/com/example/interpark/fragments/SearchPage/SearchFragment.kt
@@ -1,10 +1,9 @@
-package com.example.interpark.fragments
+package com.example.interpark.fragments.SearchPage
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
 import android.widget.EditText
 import android.widget.ImageButton
 import android.widget.Toast
