+++ app/src/main/java/com/example/interpark/fragments/SearchPage/SearchFragmentHome.kt
@@ -1,0 +1,12 @@
+package com.example.interpark.fragments.SearchPage
+
+import android.os.Bundle
+import android.view.View
+import androidx.fragment.app.Fragment
+import com.example.interpark.R
+
+class SearchFragmentHome: Fragment(R.layout.fragment_search_home) {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+    super.onViewCreated(view, savedInstanceState)
+}
+}
