+++ app/src/main/java/com/example/interpark/fragments/bookFlow/CalendarFragment.kt
@@ -1,18 +1,11 @@
-package com.example.interpark.fragments
+package com.example.interpark.fragments.bookFlow
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import android.widget.CalendarView
-import android.widget.ImageView
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.navigation.findNavController
 import androidx.navigation.fragment.findNavController
-import androidx.navigation.fragment.navArgs
-import com.example.interpark.R
 import com.example.interpark.databinding.FragmentCalendarBinding
 import java.text.SimpleDateFormat
 import java.util.*
