+++ app/src/main/java/com/example/interpark/fragments/bookFlow/PaymentFragment.kt
@@ -1,0 +1,4 @@
+package com.example.interpark.fragments.bookFlow
+
+class PaymentFragment {
+}
