+++ app/src/main/java/com/example/interpark/fragments/bookFlow/TimeSelectionFragment.kt
@@ -1,4 +1,4 @@
-package com.example.interpark.fragments
+package com.example.interpark.fragments.bookFlow
 
 import android.os.Bundle
 import android.view.LayoutInflater
